nse-intelligence/requirements.txt
fastapi
uvicorn
pandas
httpx
beautifulsoup4
lxml
python-dotenv
plotly
yfinance
